Main notebook for WB1 ICs in CliMT
Main .ipynb notebook where work is conducted in passing WB1 data as digestible ICs for CliMT.

Code is a modified version of an example piece from the CliMT GitHub page of the same name. Will need to assess which aspects are needed to fit the requirements/restrictions of a well defined DDC. Will check this with CW and CM.

Currently the code fails due to mismatched grid points (my understanding is that despite me initialising the grid before starting simulations, at some point this is re-initialised with difference lat/long coords and different levels).